Refactor role creation to support permissions and description
Updated backend and frontend to allow creating roles with associated permissions and an optional description. Backend DTOs and entities were modified to include description and permissions, and Swagger decorators were added. The frontend role creation page now supports selecting permissions and entering a description, using a new action to send the data to the backend.